requirements.txt
accelerate==1.4.0
apex==0.9.10dev
bddl==1.0.1
boto3==1.37.6
botocore==1.37.6
cloudpickle==2.1.0
fsspec==2024.12.0
datasets==3.3.2
deepdiff==8.2.0
diffusers==0.32.2
draccus==0.11.5
dynamixel_sdk==3.7.31
easydict==1.9
einops==0.8.1
filelock==3.18.0
#flash_attn==2.8.2
Flask==3.1.1
fsspec==2025.7.0
gym_pusht==0.1.5
gymnasium==0.29.1
h5py==3.13.0
huggingface_hub==0.31.1
hydra-core==1.3.2
imageio==2.37.0
jaxtyping==0.2.38
jax==0.6.2 # 0.7.0
jsonlines==4.0.0
loguru==0.7.3
matplotlib==3.5.3
mediapy==1.2.4
mujoco==3.3.0
natsort==8.4.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.99.5
opencv_python==4.11.0.86
orbax==0.1.9
packaging==25.0
pandas==2.3.1
peft==0.17.0
Pillow==11.3.0
pyarrow==19.0.1
pynput==1.8.0
pyrealsense2==2.56.5.9235
pyserial==3.5
pytest==8.3.5
pytorch_fid==0.3.0
PyYAML==6.0.2
PyYAML==6.0.2
pyzmq==27.0.1
qwen_vl_utils==0.0.11
Requests==2.32.4
rerun==1.0.30
robomimic==0.2.0
robosuite==1.4.0
s3transfer==0.11.4
safetensors==0.6.2
scipy==1.15.3 # 1.16.1
sentencepiece==0.2.0
setuptools==58.2.0 # 80.3.1
tensorboard==2.19.0
termcolor==3.1.0
thop==0.1.1.post2209072238
timm==1.0.19
# torch==2.6.0
# TorchCodec==0.2.1+cu124
# torchmetrics==1.8.1
# torchvision==0.21.0
tqdm==4.67.1
tqdm_loggable==0.2
transformers==4.53.0
ml_dtypes==0.5.3 #0.2.0
types_boto3_s3==1.37.0
tyro==0.9.27
wandb==0.21.1
